+++ src/main/java/com/hbm/blocks/bomb/NukeBoy.java
@@ -9,10 +9,11 @@ import com.hbm.config.BombConfig;
 import com.hbm.config.GeneralConfig;
 import com.hbm.entity.effect.EntityNukeTorex;
 import com.hbm.entity.logic.EntityNukeExplosionMK5;
+import com.hbm.explosion.nuclear.NuclearBurstContext;
+import com.hbm.explosion.nuclear.NuclearBurstResolver;
 import com.hbm.interfaces.IBomb;
 import com.hbm.main.MainRegistry;
 import com.hbm.tileentity.bomb.TileEntityNukeBoy;
-import com.hbm.util.TrackerUtil;
 
 import cpw.mods.fml.common.network.internal.FMLNetworkHandler;
 import net.minecraft.block.Block;
@@ -131,14 +132,12 @@ public class NukeBoy extends BlockContainer implements IBomb {
 			tetn.clearSlots();
 			world.playSoundEffect(x, y, z, "random.explode", 1.0f, world.rand.nextFloat() * 0.1F + 0.9F);
 
-			world.spawnEntityInWorld(EntityNukeExplosionMK5.statFac(world, BombConfig.boyRadius, x + 0.5, y + 0.5, z + 0.5));
+			double detonationX = x + 0.5D, detonationY = y + 0.5D, detonationZ = z + 0.5D;
+			NuclearBurstContext context = NuclearBurstResolver.resolve(world, detonationX, detonationY, detonationZ, BombConfig.boyRadius);
+			world.spawnEntityInWorld(EntityNukeExplosionMK5.statFac(world, context, detonationX, detonationY, detonationZ));
 			//world.spawnEntityInWorld(EntityNukeCloudSmall.statFac(world, x, y, z, BombConfig.boyRadius));
 
-			EntityNukeTorex torex = new EntityNukeTorex(world);
-			torex.setPositionAndRotation(x + 0.5, y + 1, z + 0.5, 0, 0);
-			torex.getDataWatcher().updateObject(10, 1.5F);
-			world.spawnEntityInWorld(torex);
-			TrackerUtil.setTrackingRange(world, torex, 1000);
+			EntityNukeTorex.statFac(world, detonationX, detonationY, detonationZ, context);
 		}
 		return false;
 	}